# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nova Kudos",
        "description": "Demo Nova Kudos Api",
        "version": "1.0"
    },
    "paths": {
...
        "/api/authentication/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "login",
                "description": "login",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "fa"
                    },
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "mobile"
                                ],
                                "properties": {
                                    "mobile": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "09350000000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "authentication/login-register"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "unauthorized"
                    },
                    "422": {
                        "description": "error"
                    },
                    "500": {
                        "description": "server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/authentication/check-otp": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "check otp",
                "description": "check otp",
                "operationId": "check otp",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "fa"
                    },
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "mobile",
                                    "otp_code"
                                ],
                                "properties": {
                                    "mobile": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "09350000000"
                                    },
                                    "otp_code": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "0912"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "/api/authentication/check-otp"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "request error"
                    },
                    "401": {
                        "description": "unauthorized"
                    },
                    "500": {
                        "description": "server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/authentication/resend-otp": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "resend otp code",
                "description": "resend otp code",
                "operationId": "resend otp code",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "fa"
                    },
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "mobile"
                                ],
                                "properties": {
                                    "mobile": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "09350000000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "/api/authentication/resend-otp"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "unauthorized"
                    },
                    "422": {
                        "description": "error"
                    },
                    "500": {
                        "description": "server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/authentication/logout": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "logout",
                "description": "logout",
                "operationId": "logout",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "fa"
                    },
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "/api/authentication/logout"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "unauthorized"
                    },
                    "422": {
                        "description": "error"
                    },
                    "500": {
                        "description": "server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}